+++ rollup.cjs.js
@@ -1,22 +1,8 @@
 
 export default {
   input: {
-    main: 'dist/main.js',
-    vdom: 'dist/vdom.js'
+    main: 'dist/main.js'
   },
-  plugins: [
-    {
-      resolveId(id, importer) {
-        if (importer) {
-          if (id === './vdom') {
-            return { id: './vdom.cjs', external: true }
-          } else {
-            return { id, external: true }
-          }
-        }
-      }
-    }
-  ],
   output: {
     format: 'cjs',
     dir: 'dist',